join websocket server
