+++ one-to-one-video/.env.local.sample
@@ -5,9 +5,8 @@ EMAIL_SERVER=smtp://username:password@smtp.example.com:587
 EMAIL_FROM=noreply@example.com
 DATABASE_URL=sqlite://localhost/:memory:?synchronize=true
 NEXTAUTH_URL=http://localhost:3000
-NEXT_API_HOST=http://localhost
 NEXT_API_PORT=3000
 NEXT_PUBLIC_TELNYX_SIP_SUBDOMAIN=example-subdomain
 NEXT_PUBLIC_TELNYX_SIP_DOMAIN=sip.telnyx.com
-NEXT_PUBLIC_WS_SERVER_URL=
-NEXT_PUBLIC_WS_SERVER_PORT=8080
+NEXT_PUBLIC_WS_SERVER_HOST=localhost:3000
+NEXT_PUBLIC_WS_SERVER_PATH=ws
