+++ one-to-one-video/package.json
@@ -3,12 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"npm:dev:*\"",
-    "dev:next": "next dev",
-    "dev:wss": "NODE_ENV=local nodemon ./wsServer.js",
+    "dev": "LOCAL_ENV=local nodemon --ignore components --ignore utils --ignore public server.js",
     "build": "next build",
-    "start": "next start",
-    "start:wss": "NODE_ENV=production.local node ./wsServer.js",
+    "start": "NODE_ENV=production node server.js",
     "setup": "npm install && cp .env.local.sample .env.local"
   },
   "dependencies": {
