+++ one-to-one-video/pages/index.js
@@ -1,79 +1,9 @@
-import React, { useMemo, useState, Fragment } from 'react';
+import React from 'react';
 import { useSession } from 'next-auth/client';
-import { TelnyxRTCProvider } from '@telnyx/react-client';
 import useCredentials from '../utils/useCredentials';
-import useWebSocket from '../utils/useWebSocket';
 import Layout from '../components/Layout';
-import VideoCall from '../components/VideoCall';
+import VideoRTC from '../components/VideoRTC';
 import SignIn from '../components/SignIn';
-
-function AuthenticatedContent({ userEmail, credentials }) {
-  const { isReady, message, sendMessage } = useWebSocket();
-  const [displayName, setDisplayName] = useState('anonymous');
-
-  // Memoize TelnyxRTC credential to prevent re-instantiating client
-  const telnyxRTCCredential = useMemo(() => {
-    return {
-      login_token: credentials.login_token,
-    };
-  }, [credentials.login_token]);
-
-  const onTelnyxReady = () => {
-    sendMessage(
-      JSON.stringify({
-        notify_clients: true,
-        status: 'user_rtc_ready',
-        user_email: userEmail,
-        sip_username: credentials.sip_username,
-      })
-    );
-  };
-
-  const onDial = (invitedEmail) => {
-    sendMessage(
-      JSON.stringify({
-        notify_clients: true,
-        status: 'user_initiated_dial',
-        user_email: userEmail,
-        destination_user_email: invitedEmail,
-      })
-    );
-  };
-
-  const onDisplayNameChange = (value) => {
-    setDisplayName(value);
-
-    sendMessage(
-      JSON.stringify({
-        notify_clients: true,
-        status: 'user_name_change',
-        user_email: userEmail,
-        display_name: value,
-      })
-    );
-  };
-
-  return (
-    <TelnyxRTCProvider
-      credential={telnyxRTCCredential}
-      // TODO remove dev
-      options={{ env: 'development' }}
-    >
-      <Fragment>
-        {isReady && (
-          <VideoCall
-            userEmail={userEmail}
-            displayName={displayName}
-            serverMessage={message}
-            onTelnyxReady={onTelnyxReady}
-            onDial={onDial}
-            onDisplayNameChange={onDisplayNameChange}
-          />
-        )}
-      </Fragment>
-    </TelnyxRTCProvider>
-  );
-}
 
 export default function Home() {
   const [session, loading] = useSession();
@@ -83,10 +13,7 @@ export default function Home() {
   return (
     <Layout title="Home">
       {isSessionReady && credentials && (
-        <AuthenticatedContent
-          userEmail={session.user.email}
-          credentials={credentials}
-        />
+        <VideoRTC userEmail={session.user.email} credentials={credentials} />
       )}
       {!isSessionReady && !loading && <SignIn />}
     </Layout>
